requirements.txt
black==24.3.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
greenlet==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
